fix - various errors

+++ Sources/HEImagePicker/View/Gallery/HEPreviewBoxView.swift
@@ -12,6 +12,7 @@ import HECommon
 import Photos
 
 public protocol HEPreviewBoxViewDelegate: AnyObject {
+    func previewBoxViewDefaultSelection(_ box: HEPreviewBoxView) -> HELibrarySelection?
     func previewBoxViewItems(_ box: HEPreviewBoxView) -> [HELibrarySelection]
     func previewBoxViewStartedLoadingImage(_ box: HEPreviewBoxView)
     func previewBoxViewFinishedLoadingImage(_ box: HEPreviewBoxView)
@@ -32,7 +33,7 @@ public class HEPreviewBoxView: UIView {
     
     public private(set) var editButton: HECapsuleButton?
     public var currentZoomableView: HEAssetZoomableView? {
-        if (delegate?.previewBoxViewItems(self).isEmpty ?? true) {
+        if items().isEmpty {
             return nil
         }
         if let cell = collView.cellForItem(at: IndexPath(row: currentIndex, section: 0)) as? HEPreviewCell {
@@ -44,16 +45,12 @@ public class HEPreviewBoxView: UIView {
     public var usingClop = PickerConfig.library.usingClop
     public var isShown = true
     public var spinnerIsShown = false
-    private var currentIndex: Int = 0 {
-        didSet {
-            trace(currentIndex)
-        }
-    }
+    private var currentIndex: Int = 0
     
-    internal var collView: UICollectionView!
+    private(set) internal var collView: UICollectionView!
     private let spinner = UIActivityIndicatorView(style: .medium)
     private var isMultipleSelectionEnabled = false
-
+    
     private var shouldShowLoader = false {
         didSet {
             DispatchQueue.main.async {
@@ -62,6 +59,15 @@ public class HEPreviewBoxView: UIView {
         }
     }
     
+    private let emptyImageView: UIImageView = {
+        let iv = UIImageView()
+        iv.image = PickerConfig.icons.emptyPhotoIcon
+        iv.sizeToFit()
+        return iv
+    }()
+    
+    /// 미디어 선택없이 미리보기 처리를 위한 플래그
+    var lastPreviwingSelection: HELibrarySelection?
     
     deinit {
         trace()
@@ -128,7 +134,7 @@ public class HEPreviewBoxView: UIView {
         let items = items()
         if let item = items.get(at: currentIndex) {
             delegate?.previewBoxViewEditButtonTouched(self, selection: item)
-        } else if let first = items.first {
+        } else if let first = lastPreviwingSelection {
             delegate?.previewBoxViewEditButtonTouched(self, selection: first)
         }
     }
@@ -156,13 +162,13 @@ public class HEPreviewBoxView: UIView {
     @objc
     func showEditButtonIfNeed(isVideoMode: Bool) {
         NSObject.cancelPreviousPerformRequests(withTarget: self, selector: #selector(showEditButtonIfNeed), object: nil)
-        guard let button = editButton, button.isHidden else {
+        guard let button = editButton, (button.isHidden || button.alpha < 1) else {
             return
         }
         guard let currentZoomableView else {
             return
         }
-        
+        trace()
         if currentZoomableView.isVideoMode {
             button.isHidden = true
             // TODO: editing video
@@ -172,7 +178,8 @@ public class HEPreviewBoxView: UIView {
         button.alpha = 0
         button.sizeToFit()
         button.transform = CGAffineTransform(translationX: 0, y: 10)
-        UIView.animate(withDuration: 0.24, delay: 0.0, options: [.curveEaseOut], animations: {
+        button.layer.removeAllAnimations()
+        UIView.animate(withDuration: 0.24, delay: 0.1, options: [.curveEaseOut, .beginFromCurrentState], animations: {
             button.alpha = 1
             button.transform = .identity
         })
@@ -180,11 +187,13 @@ public class HEPreviewBoxView: UIView {
     
     @objc
     func hideEditButton() {
+        trace()
         NSObject.cancelPreviousPerformRequests(withTarget: self, selector: #selector(hideEditButton), object: nil)
         guard let button = editButton, !button.isHidden else {
             return
         }
-        UIView.animate(withDuration: 0.24, delay: 0, animations: {
+        button.layer.removeAllAnimations()
+        UIView.animate(withDuration: 0.18, delay: 0, options: [.beginFromCurrentState], animations: {
             button.alpha = 0
         }, completion: { _ in
             button.isHidden = true
@@ -195,9 +204,12 @@ public class HEPreviewBoxView: UIView {
         let point = self.convert(collView.center, to: collView)
         if let indexPath = collView.indexPathForItem(at: point) {
             currentIndex = indexPath.row
-            if let _ = editButton,
-               let cell = collView.cellForItem(at: indexPath) as? HEPreviewCell {
+            if let cell = collView.cellForItem(at: indexPath) as? HEPreviewCell {
                 self.perform(#selector(self.showEditButtonIfNeed), with: cell.zoomableView.currentAssetType == .video, afterDelay: 0.2)
+            } 
+            else if let item = items().get(at: indexPath.row),
+                        let asset = PHAsset.fetchAssets(withLocalIdentifiers: [item.assetIdentifier], options: nil).firstObject {
+                self.perform(#selector(self.showEditButtonIfNeed), with: asset.mediaType == .video, afterDelay: 0.2)
             }
         } else {
             currentIndex = 0
@@ -259,13 +271,15 @@ extension HEPreviewBoxView {
     }
     
     private func loadPreview(_ asset: PHAsset?, forCell cell: HEPreviewCell, selection: HELibrarySelection) -> Task<(), Never> {
-        Task {
+        
+        let isNotSquare = collView.collectionViewLayout is CenteredCellFlowLayout
+        
+        return Task {
             guard let asset = asset else {
                 print("No asset to change.")
                 return
             }
             delegate?.previewBoxViewStartedLoadingImage(self)
-            let isNotSquare = collView.collectionViewLayout is CenteredCellFlowLayout
             let completion = { [weak self] (isLowResIntermediaryImage: Bool) in
                 guard let self else { return }
                 if Task.isCancelled { return }
@@ -320,7 +334,13 @@ extension HEPreviewBoxView {
     
     
     func items() -> [HELibrarySelection] {
-        return self.delegate?.previewBoxViewItems(self) ?? []
+        let all = self.delegate?.previewBoxViewItems(self) ?? []
+        if all.isEmpty {
+            if let it = delegate?.previewBoxViewDefaultSelection(self) { // 기본 미리보기
+                return [it]
+            }
+        }
+        return all
     }
     
     func setupCollView() {
@@ -334,13 +354,14 @@ extension HEPreviewBoxView {
         addSubview(coll)
         coll.translatesAutoresizingMaskIntoConstraints = false
         coll.edgesConstraintToSuperview(edges: .all)
+        coll.backgroundColor = .lightGray
         
         collView = coll
         coll.delegate = self
         coll.dataSource = self
     }
     
-    func changePreviewLayoutIfNeed() {
+    private func changePreviewLayoutIfNeed() {
         let count = items().count
         if count > 1 {
             if !(collView.collectionViewLayout is CenteredCellFlowLayout) {
@@ -361,16 +382,19 @@ extension HEPreviewBoxView {
         }
         
         if count == 0 {
-            collView.backgroundColor = .lightGray
+            collView.backgroundColor = UIColor(white: 238 / 255.0, alpha: 1.0)
+            emptyImageView.isHidden = false
+            emptyImageView.center = CGPoint(x: self.bounds.width / 2, y: self.bounds.height / 2)
+            addSubview(emptyImageView)
         } else {
             collView.backgroundColor = .white
+            emptyImageView.isHidden = false
+            emptyImageView.removeFromSuperview()
         }
         
-        if count == 1 {
+        DispatchQueue.main.asyncAfter(deadline: .now() + 0.1) { [weak self] in
+            self?.checkEditButtonShowing()
             
-            DispatchQueue.main.asyncAfter(deadline: .now() + 0.1) { [weak self] in
-                self?.checkEditButtonShowing()
-            }
         }
     }
     
@@ -392,10 +416,16 @@ extension HEPreviewBoxView {
         }
     }
     
+    func isDefaultPreviewing() -> Bool {
+        lastPreviwingSelection?.isJustPreviewing == true
+    }
+    
     func removed(at index: Int) {
-        hideEditButton()
+        
+        self.perform(#selector(self.hideEditButton), with: nil)
+        
         let items = items()
-        if index <= items.count {
+        if index <= items.count && !isDefaultPreviewing() { // dataSource 가 먼저 값이 변경되어 <= 비교
             collView.deleteItems(at: [IndexPath(row: index, section: 0)])
             if index - 1 >= 0 && items.count > 0 {
                 collView.scrollToItem(at: IndexPath(row: index - 1, section: 0), at: .centeredHorizontally, animated: true)
@@ -409,9 +439,10 @@ extension HEPreviewBoxView {
     }
     
     func inserted(at index: Int) {
-        hideEditButton()
+        self.perform(#selector(self.hideEditButton), with: nil)
+        
         let items = items()
-        if index < items.count {
+        if index < items.count && !isDefaultPreviewing() {
             let indexPath = IndexPath(row: index, section: 0)
             collView.insertItems(at: [indexPath])
             collView.scrollToItem(at: indexPath, at: .centeredHorizontally, animated: true)
@@ -425,7 +456,7 @@ extension HEPreviewBoxView {
     
     func reload(at index: Int) {
         let items = items()
-        if index < items.count {
+        if index < items.count && !items.isEmpty {
             collView.reloadItems(at: [IndexPath(row: index, section: 0)])
         } else {
             collView.reloadData()
@@ -454,12 +485,13 @@ extension HEPreviewBoxView: UIGestureRecognizerDelegate {
 extension HEPreviewBoxView: UICollectionViewDelegateFlowLayout, UICollectionViewDataSource {
     
     public func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        return items().count
+        items().count
     }
     
     public func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        let items = self.items()
         let cell = collectionView.dequeueReusableCell(withReuseIdentifier: HEPreviewCell.reuseIdentifier, for: indexPath) as! HEPreviewCell
+        
+        let items = self.items()
         cell.isZoomable = items.count < 2
         cell.usingClop = self.usingClop
         if items.count < 2 {
@@ -470,12 +502,33 @@ extension HEPreviewBoxView: UICollectionViewDelegateFlowLayout, UICollectionView
             cell.contentView.layer.masksToBounds = true
         }
         
-        if let item = items.get(at: indexPath.row) {
+        
+        return cell
+    }
+    
+    public func collectionView(_ collectionView: UICollectionView, willDisplay cell: UICollectionViewCell, forItemAt indexPath: IndexPath) {
+        let cell = cell as! HEPreviewCell
+        if let item = self.items().get(at: indexPath.row) {
+            lastPreviwingSelection = item
+            
+            if cell.bindingIdentifier == item.assetIdentifier {
+                if let hei = self.editImageStore?.getHEImage(forId: item.assetIdentifier) {
+                    if hei.updatedTime == cell.bindingTime {
+                        return
+                    }
+                } else {
+                    return
+                }
+            }
+            cell.bindingIdentifier = item.assetIdentifier
+            
             let task = Task {
                 trace("미리보기 로드")
                 if let hei = self.editImageStore?.getHEImage(forId: item.assetIdentifier) {
+                    cell.bindingTime = hei.updatedTime
                     await loadPreviewWithHEImage(hei, forCell: cell, selection: item).value
                 } else if let asset = PHAsset.fetchAssets(withLocalIdentifiers: [item.assetIdentifier], options: PHFetchOptions()).firstObject {
+                    cell.bindingTime = asset.creationDate?.timeIntervalSince1970 ?? Date().timeIntervalSince1970
                     await loadPreview(asset, forCell: cell, selection: item).value
                 } else {
                     woops("뭐지?")
@@ -493,7 +546,6 @@ extension HEPreviewBoxView: UICollectionViewDelegateFlowLayout, UICollectionView
             cell.loadTask?.cancel()
             cell.loadTask = task
         }
-        return cell
     }
     
     public func collectionView(_ collectionView: UICollectionView, didEndDisplaying cell: UICollectionViewCell, forItemAt indexPath: IndexPath) {
@@ -524,7 +576,12 @@ extension HEPreviewBoxView: UICollectionViewDelegateFlowLayout, UICollectionView
 
 class HEPreviewCell: UICollectionViewCell {
     static var reuseIdentifier = "he.PreviewCell"
+    
+    var bindingIdentifier: String?
+    var bindingTime: TimeInterval?
+    
     lazy var zoomableView = HEAssetZoomableView()
+    
     var squareCropButton: UIButton?
     
     var isZoomable: Bool = true {
